Move GetShapedTextPath to SKPaintExtensions, cache SKPaint.GetFont reflection
