+++ TwitchDownloaderCore/ChatRenderer.cs
@@ -1191,7 +1191,7 @@ namespace TwitchDownloaderCore
                 if (renderOptions.Outline)
                 {
                     using var outlinePath = isRtl
-                        ? GetShapedTextPath(textFont, drawText, drawPos.X, drawPos.Y)
+                        ? textFont.GetShapedTextPath(drawText, drawPos.X, drawPos.Y)
                         : textFont.GetTextPath(drawText, drawPos.X, drawPos.Y);
 
                     sectionImageCanvas.DrawPath(outlinePath, outlinePaint);
@@ -1297,52 +1297,6 @@ namespace TwitchDownloaderCore
             buffer.Add(rtlText, textFont.TextEncoding);
             SKShaper.Result measure = shaper.Shape(buffer, textFont);
             return measure.Width;
-        }
-
-        // Heavily modified from SkiaSharp.HarfBuzz.CanvasExtensions.DrawShapedText
-        private static SKPath GetShapedTextPath(SKPaint paint, ReadOnlySpan<char> text, float x, float y)
-        {
-            var returnPath = new SKPath();
-
-            if (text.IsEmpty || text.IsWhiteSpace())
-                return returnPath;
-
-            using var shaper = new SKShaper(paint.Typeface);
-            using var buffer = new HarfBuzzSharp.Buffer();
-            buffer.Add(text, paint.TextEncoding);
-            var result = shaper.Shape(buffer, x, y, paint);
-
-            var glyphSpan = result.Codepoints.AsSpan();
-            var pointSpan = result.Points.AsSpan();
-
-            var xOffset = 0.0f;
-            if (paint.TextAlign != SKTextAlign.Left)
-            {
-                var width = result.Width;
-                if (paint.TextAlign == SKTextAlign.Center)
-                    width *= 0.5f;
-                xOffset -= width;
-            }
-
-            // We cannot dispose because it is a reference, not a clone.
-            var getFont = typeof(SKPaint).GetMethod("GetFont", BindingFlags.NonPublic | BindingFlags.Instance)!;
-            var font = (SKFont)getFont.Invoke(paint, Array.Empty<object>())!;
-            for (var i = 0; i < pointSpan.Length; i++)
-            {
-                using var glyphPath = font.GetGlyphPath((ushort)glyphSpan[i]);
-                if (glyphPath.IsEmpty)
-                    continue;
-
-                var point = pointSpan[i];
-                glyphPath.Transform(new SKMatrix(
-                    1, 0, point.X + xOffset,
-                    0, 1, point.Y,
-                    0, 0, 1
-                ));
-                returnPath.AddPath(glyphPath);
-            }
-
-            return returnPath;
         }
 
         private void DrawUsername(Comment comment, List<(SKImageInfo info, SKBitmap bitmap)> sectionImages, ref Point drawPos, Point defaultPos, bool appendColon = true, string colorOverride = null)
